+++ components/browser/state/src/test/java/mozilla/components/browser/state/store/BrowserStoreExceptionTest.kt
@@ -5,11 +5,8 @@
 package mozilla.components.browser.state.store
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import mozilla.components.browser.state.action.TabGroupAction
 import mozilla.components.browser.state.action.TabListAction
 import mozilla.components.browser.state.state.BrowserState
-import mozilla.components.browser.state.state.TabGroup
-import mozilla.components.browser.state.state.TabPartition
 import mozilla.components.browser.state.state.createTab
 import mozilla.components.browser.state.state.recover.toRecoverableTab
 import org.junit.Test
@@ -79,80 +76,5 @@ class BrowserStoreExceptionTest {
             )
 
         store.dispatch(TabListAction.AddMultipleTabsAction(tabs = listOf(tab1, tab2)))
-    }
-
-    @Test(expected = IllegalArgumentException::class)
-    fun `AddTabGroupAction - Exception is thrown when group already exists`() {
-        val partitionId = "testFeaturePartition"
-        val testGroup = TabGroup("test")
-        val store =
-            BrowserStore(
-                BrowserState(
-                    tabPartitions =
-                        mapOf(
-                            partitionId to
-                                TabPartition(
-                                    partitionId,
-                                    tabGroups = listOf(testGroup),
-                                )
-                        )
-                )
-            )
-
-        store.dispatch(
-            TabGroupAction.AddTabGroupAction(
-                partition = partitionId,
-                group = testGroup,
-            )
-        )
-    }
-
-    @Test(expected = IllegalArgumentException::class)
-    fun `AddTabGroupAction - Asserts that tabs exist`() {
-        val store = BrowserStore()
-
-        val partition = "testFeaturePartition"
-        val testGroup = TabGroup("test", tabIds = setOf("invalid"))
-        store.dispatch(
-            TabGroupAction.AddTabGroupAction(
-                partition = partition,
-                group = testGroup,
-            )
-        )
-    }
-
-    @Test(expected = IllegalArgumentException::class)
-    fun `AddTabAction - Asserts that tab exists when adding to group`() {
-        val tabGroup = TabGroup("test1", tabIds = emptySet())
-        val tabPartition = TabPartition("testFeaturePartition", tabGroups = listOf(tabGroup))
-
-        val store =
-            BrowserStore(
-                BrowserState(
-                    tabs = listOf(),
-                    tabPartitions = mapOf("testFeaturePartition" to tabPartition),
-                )
-            )
-
-        val tab = createTab(id = "tab1", url = "https://firefox.com")
-        store.dispatch(TabGroupAction.AddTabAction(tabPartition.id, tabGroup.id, tab.id))
-    }
-
-    @Test(expected = IllegalArgumentException::class)
-    fun `AddTabsAction - Asserts that tabs exist when adding to group`() {
-        val tabGroup = TabGroup("test1", tabIds = emptySet())
-        val tabPartition = TabPartition("testFeaturePartition", tabGroups = listOf(tabGroup))
-        val tab1 = createTab(id = "tab1", url = "https://firefox.com")
-        val tab2 = createTab(id = "tab2", url = "https://mozilla.org")
-
-        val store =
-            BrowserStore(
-                BrowserState(
-                    tabs = listOf(tab1),
-                    tabPartitions = mapOf("testFeaturePartition" to tabPartition),
-                )
-            )
-
-        store.dispatch(TabGroupAction.AddTabsAction(tabPartition.id, tabGroup.id, setOf(tab1.id, tab2.id)))
     }
 }