+++ components/browser/tabstray/src/test/java/mozilla/components/browser/tabstray/TabsAdapterTest.kt
@@ -72,7 +72,6 @@ class TabsAdapterTest {
                 createTab(id = "A", url = "https://www.mozilla.org"),
                 createTab(id = "B", url = "https://www.firefox.com"),
             ),
-            tabPartition = null,
             selectedTabId = "A",
         )
         assertEquals(2, adapter.itemCount)
@@ -90,8 +89,7 @@ class TabsAdapterTest {
 
         adapter.updateTabs(
             listOf(tab),
-            null,
-            "A",
+            tab.id,
         )
 
         adapter.onBindViewHolder(holder, 0)
@@ -105,7 +103,7 @@ class TabsAdapterTest {
         val holder = spy(TestTabViewHolder(View(testContext)))
         val tab = createTab(id = "A", url = "https://www.mozilla.org")
 
-        adapter.updateTabs(listOf(mock(), tab), tabPartition = null, selectedTabId = "A")
+        adapter.updateTabs(listOf(mock(), tab), selectedTabId = "A")
 
         adapter.onBindViewHolder(holder, 0, listOf(PAYLOAD_HIGHLIGHT_SELECTED_ITEM))
         verify(holder, never()).updateSelectedTabIndicator(ArgumentMatchers.anyBoolean())
@@ -119,7 +117,7 @@ class TabsAdapterTest {
         val adapter = TabsAdapter(delegate = mock())
         val holder = spy(TestTabViewHolder(View(testContext)))
         val tab = createTab(id = "A", url = "https://www.mozilla.org")
-        adapter.updateTabs(listOf(mock(), tab), tabPartition = null, selectedTabId = "A")
+        adapter.updateTabs(listOf(mock(), tab), selectedTabId = "A")
 
         adapter.onBindViewHolder(holder, 0, listOf(PAYLOAD_DONT_HIGHLIGHT_SELECTED_ITEM))
         verify(holder, never()).updateSelectedTabIndicator(ArgumentMatchers.anyBoolean())
@@ -139,7 +137,7 @@ class TabsAdapterTest {
         verify(payloads, never()).isEmpty()
         verify(payloads, never()).contains(ArgumentMatchers.anyInt())
 
-        adapter.updateTabs(emptyList(), tabPartition = null, selectedTabId = null)
+        adapter.updateTabs(emptyList(), selectedTabId = null)
         adapter.onBindViewHolder(holder, 0, payloads)
         // verify that calls we expect further down are not happening after the null check
         verify(payloads, never()).isEmpty()
@@ -152,7 +150,7 @@ class TabsAdapterTest {
         val holder = TestTabViewHolder(View(testContext))
         val emptyPayloads = spy(arrayListOf<String>())
 
-        adapter.updateTabs(listOf(mock()), tabPartition = null, selectedTabId = null)
+        adapter.updateTabs(listOf(mock()), selectedTabId = null)
 
         adapter.onBindViewHolder(holder, 0, emptyPayloads)