+++ components/feature/session/src/main/java/mozilla/components/feature/session/middleware/undo/UndoMiddleware.kt
@@ -133,8 +133,7 @@ class UndoMiddleware(
 
         val undoHistory = state.undoHistory
         val tabs = undoHistory.tabs
-        val tabPartitions = undoHistory.tabPartitions
-        if (tabs.isEmpty() && tabPartitions.isEmpty()) {
+        if (tabs.isEmpty()) {
             logger.debug("No recoverable tabs or tab partitions for undo.")
             return@launch
         }
@@ -143,7 +142,6 @@ class UndoMiddleware(
             TabListAction.RestoreAction(
                 tabs = tabs,
                 restoreLocation = TabListAction.RestoreAction.RestoreLocation.AT_INDEX,
-                tabPartitions = tabPartitions,
             )
         )