+++ components/feature/tabs/src/test/java/mozilla/components/feature/tabs/TabsUseCasesTest.kt
@@ -17,11 +17,8 @@ import mozilla.components.browser.state.engine.EngineMiddleware
 import mozilla.components.browser.state.selector.findNormalOrPrivateTabByUrl
 import mozilla.components.browser.state.selector.findTab
 import mozilla.components.browser.state.selector.selectedTab
-import mozilla.components.browser.state.state.TabGroup
-import mozilla.components.browser.state.state.TabPartition
 import mozilla.components.browser.state.state.TabSessionState
 import mozilla.components.browser.state.state.createTab
-import mozilla.components.browser.state.state.getGroupById
 import mozilla.components.browser.state.state.recover.RecoverableTab
 import mozilla.components.browser.state.state.recover.toRecoverableTab
 import mozilla.components.browser.state.store.BrowserStore
@@ -30,7 +27,6 @@ import mozilla.components.concept.engine.EngineSession
 import mozilla.components.concept.engine.EngineSession.LoadUrlFlags
 import mozilla.components.concept.engine.EngineSessionState
 import mozilla.components.concept.storage.HistoryMetadataKey
-import mozilla.components.feature.tabs.ext.tabGroupsPartition
 import mozilla.components.support.test.any
 import mozilla.components.support.test.argumentCaptor
 import mozilla.components.support.test.mock
@@ -406,7 +402,6 @@ class TabsUseCasesTest {
                 RecoverableBrowserState(
                     tabs = restoredTabs.map { it.toRecoverableTab() },
                     selectedTabId = null,
-                    tabPartitions = emptyMap(),
                 )
             val sessionStorage: SessionStorage = mock()
             whenever(sessionStorage.restore(any())).thenReturn(recoverableBrowserState)
@@ -423,21 +418,18 @@ class TabsUseCasesTest {
         }
 
     @Test
-    fun `GIVEN a recoverable browser state with tabs and partitions in storage WHEN browsing session is restored THEN restore the tabs and partition from storage`() =
+    fun `GIVEN a recoverable browser state with tabs in storage WHEN browsing session is restored THEN restore the tabs from storage`() =
         runTest(testDispatcher) {
             val restoredTabs =
                 listOf(
                     createTab(id = "tab1", url = "https://mozilla.org"),
                     createTab(id = "tab2", url = "https://firefox.com"),
                 )
-            val tabGroup = TabGroup("group1", tabIds = setOf("tab1"))
-            val tabPartition = TabPartition("testFeaturePartition", tabGroups = listOf(tabGroup))
-            val restoredTabPartitions = mapOf("testFeaturePartition" to tabPartition)
+
             val recoverableBrowserState =
                 RecoverableBrowserState(
                     tabs = restoredTabs.map { it.toRecoverableTab() },
                     selectedTabId = null,
-                    tabPartitions = restoredTabPartitions,
                 )
             val sessionStorage: SessionStorage = mock()
             whenever(sessionStorage.restore(any())).thenReturn(recoverableBrowserState)
@@ -449,7 +441,6 @@ class TabsUseCasesTest {
                 assertEquals(restoredTab.id, store.state.tabs[index].id)
                 assertEquals(restoredTab.content.url, store.state.tabs[index].content.url)
             }
-            assertEquals(restoredTabPartitions, store.state.tabPartitions)
         }
 
     @Test
@@ -460,7 +451,6 @@ class TabsUseCasesTest {
                 RecoverableBrowserState(
                     tabs = restoredTabs.map { it.toRecoverableTab() },
                     selectedTabId = null,
-                    tabPartitions = emptyMap(),
                     isTranslationsEngineSupported = true,
                 )
             val sessionStorage: SessionStorage = mock()
@@ -479,7 +469,6 @@ class TabsUseCasesTest {
                 RecoverableBrowserState(
                     tabs = restoredTabs.map { it.toRecoverableTab() },
                     selectedTabId = null,
-                    tabPartitions = emptyMap(),
                     isTranslationsEngineSupported = null,
                 )
             val sessionStorage: SessionStorage = mock()
@@ -764,95 +753,5 @@ class TabsUseCasesTest {
         assertThrows(IllegalStateException::class.java) {
             tabsUseCases.migratePrivateTabUseCase("invalid-tab-id")
         }
-    }
-
-    @Test
-    fun `WHEN AddTabGroupUseCase is invoked THEN group is added to the tab groups partition`() {
-        val tab = createTab("https://mozilla.org")
-        store.dispatch(TabListAction.AddTabAction(tab))
-
-        val group = TabGroup(id = "group1", name = "Group 1", tabIds = setOf(tab.id))
-        tabsUseCases.addTabGroup(group = group)
-
-        val partition = store.state.tabGroupsPartition()
-        assertNotNull(partition)
-        assertEquals(TabPartitionKeys.TAB_GROUPS, partition.id)
-        assertEquals(1, partition.tabGroups.size)
-        assertEquals(group, partition.getGroupById("group1"))
-    }
-
-    @Test
-    fun `WHEN CloseTabGroupUseCase is invoked THEN group and tabs are removed from the tab groups partition`() {
-        val tab = createTab("https://mozilla.org")
-        store.dispatch(TabListAction.AddTabAction(tab))
-
-        val group = TabGroup(id = "group1", name = "Group 1", tabIds = setOf(tab.id))
-        tabsUseCases.addTabGroup(group = group)
-
-        assertEquals(1, store.state.tabs.size)
-        assertEquals(1, store.state.tabGroupsPartition()?.tabGroups?.size)
-        assertEquals(group, store.state.tabGroupsPartition()?.tabGroups?.first())
-
-        tabsUseCases.closeTabGroup(
-            group = "group1",
-            tabIds = listOf(tab.id),
-        )
-
-        assertEquals(0, store.state.tabs.size)
-        assertNull(store.state.tabGroupsPartition())
-    }
-
-    @Test
-    fun `WHEN RemoveTabGroupUseCase is invoked THEN group is removed from the tab groups partition`() {
-        val tab = createTab("https://mozilla.org")
-        store.dispatch(TabListAction.AddTabAction(tab))
-
-        val group = TabGroup(id = "group1", name = "Group 1", tabIds = setOf(tab.id))
-        tabsUseCases.addTabGroup(group = group)
-
-        assertEquals(1, store.state.tabs.size)
-        assertEquals(1, store.state.tabGroupsPartition()?.tabGroups?.size)
-        assertEquals(group, store.state.tabGroupsPartition()?.tabGroups?.first())
-
-        tabsUseCases.removeTabGroup(group = "group1")
-
-        assertEquals(1, store.state.tabs.size)
-        assertEquals(tab, store.state.tabs.first())
-        assertNull(store.state.tabGroupsPartition())
-    }
-
-    @Test
-    fun `WHEN AddTabsInGroupUseCase is invoked THEN tabs are added to group in the tab groups partition`() {
-        val tab1 = createTab("https://mozilla.org")
-        val tab2 = createTab("https://firefox.com")
-        store.dispatch(TabListAction.AddMultipleTabsAction(tabs = listOf(tab1, tab2)))
-
-        tabsUseCases.addTabsInGroup(group = "group1", tabId = tab1.id)
-        var group = store.state.tabGroupsPartition()?.getGroupById("group1")
-        assertNotNull(group)
-        assertEquals(setOf(tab1.id), group.tabIds)
-
-        tabsUseCases.addTabsInGroup(group = "group1", tabIds = setOf(tab2.id))
-        group = store.state.tabGroupsPartition()?.getGroupById("group1")
-        assertEquals(setOf(tab1.id, tab2.id), group?.tabIds)
-    }
-
-    @Test
-    fun `WHEN RemoveTabsInGroupUseCase is invoked THEN tabs are removed from group in the tab groups partition`() {
-        val tab1 = createTab("https://mozilla.org")
-        val tab2 = createTab("https://firefox.com")
-        store.dispatch(TabListAction.AddMultipleTabsAction(tabs = listOf(tab1, tab2)))
-        tabsUseCases.addTabsInGroup(
-            group = "group1",
-            tabIds = setOf(tab1.id, tab2.id),
-        )
-
-        tabsUseCases.removeTabsInGroup(group = "group1", tabId = tab1.id)
-        var group = store.state.tabGroupsPartition()?.getGroupById("group1")
-        assertEquals(setOf(tab2.id), group?.tabIds)
-
-        tabsUseCases.removeTabsInGroup(group = "group1", tabIds = setOf(tab2.id))
-        group = store.state.tabGroupsPartition()?.getGroupById("group1")
-        assertTrue(group?.tabIds?.isEmpty() == true)
     }
 }