+++ components/feature/tabs/src/test/java/mozilla/components/feature/tabs/tabstray/TabsFeatureTest.kt
@@ -101,7 +101,7 @@ class TabsFeatureTest {
         tabsFeature.filterTabs(filter)
 
         verify(presenter).tabsFilter = filter
-        verify(tabsTray).updateTabs(emptyList(), null, null)
+        verify(tabsTray).updateTabs(emptyList(), null)
     }
 
     @Test