+++ components/feature/tabs/src/test/java/mozilla/components/feature/tabs/tabstray/TabsTrayPresenterTest.kt
@@ -9,7 +9,6 @@ import kotlinx.coroutines.test.StandardTestDispatcher
 import kotlinx.coroutines.test.runTest
 import mozilla.components.browser.state.action.TabListAction
 import mozilla.components.browser.state.state.BrowserState
-import mozilla.components.browser.state.state.TabPartition
 import mozilla.components.browser.state.state.TabSessionState
 import mozilla.components.browser.state.state.createTab
 import mozilla.components.browser.state.store.BrowserStore
@@ -47,7 +46,6 @@ class TabsTrayPresenterTest {
                     tabsTray,
                     store,
                     closeTabsTray = {},
-                    tabPartitionsFilter = { null },
                     tabsFilter = { true },
                     mainDispatcher = testDispatcher,
                 )
@@ -89,7 +87,6 @@ class TabsTrayPresenterTest {
                     tabsTray,
                     store,
                     closeTabsTray = {},
-                    tabPartitionsFilter = { null },
                     tabsFilter = { true },
                     mainDispatcher = testDispatcher,
                 )
@@ -130,7 +127,6 @@ class TabsTrayPresenterTest {
                     tabsTray,
                     store,
                     closeTabsTray = {},
-                    tabPartitionsFilter = { null },
                     tabsFilter = { true },
                     mainDispatcher = testDispatcher,
                 )
@@ -175,7 +171,6 @@ class TabsTrayPresenterTest {
                     tabsTray,
                     store,
                     closeTabsTray = {},
-                    tabPartitionsFilter = { null },
                     tabsFilter = { true },
                     mainDispatcher = testDispatcher,
                 )
@@ -218,7 +213,6 @@ class TabsTrayPresenterTest {
                     tabsTray,
                     store,
                     closeTabsTray = {},
-                    tabPartitionsFilter = { null },
                     tabsFilter = { true },
                     mainDispatcher = testDispatcher,
                 )
@@ -257,7 +251,6 @@ class TabsTrayPresenterTest {
                     tabsTray,
                     store,
                     closeTabsTray = {},
-                    tabPartitionsFilter = { null },
                     tabsFilter = { it.content.private },
                     mainDispatcher = testDispatcher,
                 )
@@ -293,7 +286,6 @@ class TabsTrayPresenterTest {
                 TabsTrayPresenter(
                     tabsTray,
                     store,
-                    tabPartitionsFilter = { null },
                     tabsFilter = { true },
                     closeTabsTray = { closed = true },
                     mainDispatcher = testDispatcher,
@@ -334,7 +326,6 @@ class TabsTrayPresenterTest {
                 TabsTrayPresenter(
                     tabsTray,
                     store,
-                    tabPartitionsFilter = { null },
                     tabsFilter = { true },
                     closeTabsTray = { closed = true },
                     mainDispatcher = testDispatcher,
@@ -375,7 +366,6 @@ class TabsTrayPresenterTest {
                 TabsTrayPresenter(
                     tabsTray,
                     store,
-                    tabPartitionsFilter = { null },
                     tabsFilter = { it.content.private },
                     closeTabsTray = { invoked = true },
                     mainDispatcher = testDispatcher,
@@ -392,7 +382,7 @@ private class MockedTabsTray : TabsTray {
     var updateTabs: List<TabSessionState>? = null
     var selectedTabId: String? = null
 
-    override fun updateTabs(tabs: List<TabSessionState>, tabPartition: TabPartition?, selectedTabId: String?) {
+    override fun updateTabs(tabs: List<TabSessionState>, selectedTabId: String?) {
         updateTabs = tabs
         this.selectedTabId = selectedTabId
     }